examples/requirements/test.txt
factory_boy==2.7.0
fake-factory==0.7.2
Faker==0.7.17
py==1.4.31
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-ordering==0.5
pytest==3.0.2
selenium==2.53.6
tox==2.9.1
